requirements.txt

reflex==0.8.27
langchain-groq
python-dotenv
langchain-core